LCM: implement ideas from LCM IJCAI 2017 paper
In "An Effective Learnt Clause Minimization Approach for CDCL SAT Solvers"
a learned clause simplification is described, which runs vivification on
the better half of the learned clauses once per clause, whenever a DB
reduction has been performed.

Furthermore, when vivification finds a conflict (or an positively implied
literal), the reference implementation of the paper runs conflict analysis
to reduce the clauses further.

This commit introduced a parameter "-lcm=X", that enables the above steps,
increasing the power of reasoning step by step. Furthermore, lcm-full is
added to bypass filters and run LCM after each restart on all clauses. This
helps to debug problems with the algorithm.

Currently open is the full integration in the proof generation. While DRUP
proofs should work with the current state, DRAT proofs might fail, as clause
removal is not implemented properly.