cmake: avoid clashing link names
In case of parallel compilation, the two remove commands might be
executed first, and next the two create_symlink commands might be
executed, resulting in a failing attempt, which aborts the compilation.

To avoid this problem, make sure the link names are unique. As riss is
typically not used with the default simplifier, no riss link will be
created during compilation of riss-simp.