model-counting: allow to simplify
Simplification for model counting requires to keep the number of models,
which can be achieved by (1) keeping the formula equivalent with all its
variables, (2) remove 'defined' variables, whose value is determined
by the value of other variables, like gate outputs, or (3) by introducing
variables that are defined over existing variables.

This change adds a script that allows to use Coprocessor to simplifiy
given formulas for the above (1) and (3). For (2), a few modifications
would be required, especially using the white list would have to be
dropped, and in return, techniques that would modify the number of models
like BVE or BCE, would have to be restricted.

In case of weighted counting, the number of variables has to remain either
way, because we cannot use variable densing for variables with weights,
unless we actually modify Coprocessor to consider variable weights as well.

Signed-off-by: Norbert Manthey <nmanthey@conp-solutions.com>